refactor: Breweries 컴포넌트 분리, handleClick 기능 분리 - Breweries -> Breweries, BreweryCard로 분리 - handleClick에서 잡다하게 수행하고 있는 상황 - 이전 정보를 저장하기만 하는 saveSearchInfo 함수를 BreweryCard에 전달 - BreweryCard의 handleClick은 saveSearchInfo 사용, 상세 페이지로 라우팅
