+++ client/app/search/breweries/brewery-card/BreweryCard.module.scss
@@ -1,0 +1,42 @@
+.brewery_item {
+  list-style: none;
+  padding: 14px;
+  width: 100%;
+  cursor: pointer;
+  transition: background-color 0.1s ease;
+}
+
+.brewery_item:not(:first-child) {
+  border-top: 1px solid #909090;
+  border-width: 1px;
+  border-left: 20px;
+}
+
+.brewery_item:hover {
+  background-color: rgba(80, 76, 76, 0.289);
+}
+
+.inner_box {
+  display: flex;
+  column-gap: 18px;
+  width: 100%;
+  height: 80%;
+}
+
+.brewery_desc {
+  width: 100%;
+  font-size: 14px;
+  color: #504c4c;
+  margin-top: 4px;
+  margin-bottom: 10px;
+}
+
+.recommend_title {
+  font-size: 14px;
+  color: #dd9400;
+  margin-right: 8px;
+}
+
+.address {
+  margin-top: 14px;
+}
